separate list view render and key help component
