+++ internal/ui/render.go
@@ -6,194 +6,16 @@ import (
 	"time"
 
 	"charm.land/lipgloss/v2"
-	"github.com/SayYoungMan/tfui/pkg/terraform"
 	"github.com/charmbracelet/x/ansi"
 )
 
-func (m Model) renderListView() string {
-	var s strings.Builder
-
-	fmt.Fprint(&s, m.renderFilterBox())
-	fmt.Fprintln(&s, m.renderResourcesBox())
-	fmt.Fprintln(&s, m.renderInfoBar())
-	s.WriteString("\n" + m.renderHelpBar() + "\n")
-
-	return s.String()
-}
-
-func (m Model) renderFilterBox() string {
-	var s strings.Builder
-
-	filterIcon := "⌕ "
-	filterContent := filterIcon + m.filterInput.View()
-	if m.viewState == viewFilter {
-		fmt.Fprintln(&s, focusedBorderStyle.Width(m.viewWidth).Render(filterContent))
-	} else {
-		fmt.Fprintln(&s, borderStyle.Width(m.viewWidth).Render(filterContent))
-	}
-
-	return s.String()
-}
-
-func (m Model) renderResourcesBox() string {
-	var resources strings.Builder
-	end := min(m.offset+m.visibleRows(), len(m.rows))
-	for i := m.offset; i < end; i++ {
-		row := m.rows[i]
-
-		var line string
-		if row.Item.IsResource() {
-			line = m.renderResourceLine(i)
-		} else {
-			line = m.renderModuleLine(i)
-		}
-
-		// Truncate the end to fit to screen
-		if maxLineWidth := m.viewWidth - 4; lipgloss.Width(line) > maxLineWidth {
-			line = ansi.Truncate(line, maxLineWidth, "…")
-		}
-
-		fmt.Fprintln(&resources, line)
-	}
-
-	rendered := end - m.offset
-	for range m.visibleRows() - rendered {
-		fmt.Fprintln(&resources)
-	}
-
-	renderString := strings.TrimSuffix(resources.String(), "\n")
-	return resourceBorderStyle.Width(m.viewWidth).Render(renderString)
-}
-
-func (m Model) renderResourceLine(idx int) string {
-	row := m.rows[idx]
-	addr := row.Item.Address()
-	r := m.resources[addr]
-
-	if r.Reason != "" {
-		addr += fmt.Sprintf(" (%s)", r.Reason)
-	}
-	adornment := r.Action.Symbol()
-
-	currentModule := m.currentCursorModule()
-	prefix := row.TreePrefix
-	if currentModule == row.Item.Parent.Module {
-		prefix = treePrefixCurrentStyle.Render(prefix)
-	} else {
-		prefix = treePrefixDefaultStyle.Render(prefix)
-	}
-
-	line := fmt.Sprintf("%s %s", adornment, addr)
-	switch {
-	case idx == m.cursor:
-		line = cursorStyle.Render(line)
-	case m.isSelectedOrAncestor(row.Item):
-		line = selectedStyle.Render(line)
-	}
-	if style, ok := actionStyles[r.Action]; ok {
-		line = style.Render(line)
-	}
-
-	return prefix + line
-}
-
-func (m Model) renderModuleLine(idx int) string {
-	row := m.rows[idx]
-
-	symbol := "▾"
-	if m.collapsed[row.Item.Address()] {
-		symbol = "▸"
-	}
-	line := fmt.Sprintf("%s %s", symbol, row.Item.Address())
-
-	switch {
-	case idx == m.cursor:
-		line = cursorStyle.Render(line)
-	case m.isSelectedOrAncestor(row.Item):
-		line = selectedStyle.Render(line)
-	}
-
-	prefix := row.TreePrefix
-	if m.currentCursorModule() == row.Item.Module {
-		prefix = treePrefixCurrentStyle.Render(prefix)
-		line = treePrefixCurrentStyle.Render(line)
-	} else {
-		prefix = treePrefixDefaultStyle.Render(prefix)
-		line = moduleStyle.Render(line)
-	}
-
-	return prefix + line
-}
-
-func (m Model) renderInfoBar() string {
-	var adornment, info string
-
-	switch m.workState {
-	case workStatePull:
-		adornment = infoBarStyle.Render(m.spinner.View())
-		info = " Scanning..."
-	case workPlan:
-		adornment = infoBarStyle.Render(m.spinner.View())
-		var count int
-		for _, r := range m.resources {
-			if r.Action != terraform.ActionUncertain {
-				count++
-			}
-		}
-		info = fmt.Sprintf(" Scanning... (%d/%d resources scanned)", count, len(m.resources))
-	default:
-		adornment = lipgloss.NewStyle().Foreground(colorGreen).Render("✓")
-		info = fmt.Sprintf("  Scan Complete (%d resources scanned)", len(m.resources))
-	}
-
-	if m.filterInput.Value() != "" {
-		info += fmt.Sprintf(" | showing %d", len(m.rows))
-	}
-	if len(m.selected) > 0 {
-		info += fmt.Sprintf(" | %d selected", len(m.selected))
-	}
-	if len(m.diagnostics) > 0 {
-		info += fmt.Sprintf(" | %d warnings", len(m.diagnostics))
-	}
-	return " " + adornment + infoBarStyle.Render(info)
-}
-
-func renderKeyHint(key, desc string) string {
-	key = "'" + key + "'"
-	return helpKeyStyle.Render(key) + helpDescStyle.Render(" "+desc)
-}
-
-func (m Model) renderHelpBar() string {
-	var HKeyInfo string
-	if m.hideUnchanged {
-		HKeyInfo = "show unchanged"
-	} else {
-		HKeyInfo = "hide unchanged"
-	}
-
-	hints := []string{
-		renderKeyHint("/", "filter"),
-		renderKeyHint("Space", "select"),
-		renderKeyHint("Enter", "detail"),
-		renderKeyHint("Tab", "action"),
-		renderKeyHint("H", HKeyInfo),
-		renderKeyHint("Ctrl+r", "refresh"),
-		renderKeyHint("q", "quit"),
-	}
-
-	if m.viewWidth >= 90 {
-		return " " + strings.Join(hints, "  ")
-	}
-
-	mid := (len(hints) + 1) / 2
-	line1 := " " + strings.Join(hints[:mid], "  ")
-	line2 := " " + strings.Join(hints[mid:], "  ")
-	return line1 + "\n" + line2
-}
-
 func (m Model) renderActionPickerView() string {
 	title := fmt.Sprintf("%d resource(s) selected", len(m.selected))
-	help := "Enter to choose | Esc to cancel"
+	keyInfo := []keyInfo{
+		{key: "Enter", info: "select"},
+		{key: "Esc", info: "cancel"},
+	}
+	help := m.renderKeyInfo(keyInfo)
 
 	width := max(lipgloss.Width(title), lipgloss.Width(help)) + 6
 	centered := lipgloss.NewStyle().Width(width).Align(lipgloss.Center)
@@ -431,7 +253,7 @@ func (m Model) renderDetailView() string {
 		visualRows += lineRows
 	}
 
-	box := resourceBorderStyle.Width(m.viewWidth - 2).Height(boxHeight).Render(strings.TrimSuffix(content.String(), "\n"))
+	box := borderStyle.Width(m.viewWidth - 2).Height(boxHeight).Render(strings.TrimSuffix(content.String(), "\n"))
 
 	help := "↑/↓ scroll | Esc to close"
 
@@ -457,7 +279,11 @@ func (m Model) renderQuitConfirmLayer() *lipgloss.Layer {
 	}
 	buttons := lipgloss.JoinHorizontal(lipgloss.Top, cancelButton, "  ", confirmButton)
 
-	help := "Enter to select | Esc to cancel"
+	keyInfo := []keyInfo{
+		{key: "Enter", info: "select"},
+		{key: "Esc", info: "cancel"},
+	}
+	help := m.renderKeyInfo(keyInfo)
 
 	width := lipgloss.Width(help) + 4
 	centered := lipgloss.NewStyle().Width(width).Align(lipgloss.Center)
